Refactor: Split aiChat.ts into modular files
- Split worker/routes/aiChat.ts into four focused files:
  * aiChatShared.ts - Shared utilities and constants
  * aiChatIntake.ts - Intake-specific functionality
  * aiChatOnboarding.ts - Onboarding-specific functionality
  * aiChat.ts - Main router (refactored)

- Fixed require() in aiChatIntake.ts to use proper ES imports
- Removed unused normalizeOnboardingServices import
- Removed unsupported tool_choice/parallel_tool_calls from Workers AI
- Added metadataMergeMode to ConversationService for patch/replace behavior
- All TypeScript checks pass with zero errors

Maintains exact same functionality while providing cleaner modular structure.
